requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pybit==5.6.2
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
